Add protections to python_invoke. #0
This introduces the CLASSAD_USER_PYTHON_MODULES configuration knob.
If set, the referenced python modules are loaded and made available
to the ClassAd library through the "python_invoke" function.  No
other modules are available; further, any functions starting with
"_" are disallowed.

With this, the sysadmin has fine-grained control over which
functions are available through ClassAds.