Allow Schedd.query to accept ExprTree inputs.
A key oversight of the original API is that Schedd.query only accepted
a query string (this reflected the HTCondor internals, where ExprTree
is not used for the query constraint).

This converts the argument to a ExprTree; this means simple conversions
like python 'True' becomes ClassAd 'true'.  However, to keep backward
compatibility, a string input is treated as an expression to be parsed.